+++ app/src/main/java/com/example/learningtest/fourComponents/service/bound/SimpleBoundMusicService.kt
@@ -1,4 +1,4 @@
-package com.example.learningtest.fourComponents.service
+package com.example.learningtest.fourComponents.service.bound
 
 import android.app.Service
 import android.content.Intent
