+++ app/src/main/java/com/example/learningtest/fourComponents/service/foregroundboud/Song.kt
@@ -1,4 +1,4 @@
-package com.example.learningtest.fourComponents.service
+package com.example.learningtest.fourComponents.service.foregroundboud
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.RawRes
